Revert change to mailimap_literal_data_send signature, to preserve binary compatibility with master.
Instead, cast the parameter at each callsite where size_t is used
(3 occurrences).

+++ src/low-level/imap/mailimap.c
@@ -637,7 +637,7 @@ int mailimap_append(mailimap * session, const char * mailbox,
                                                 session->imap_progress_context);
   }
   else {
-    r = mailimap_literal_data_send(session->imap_stream, literal, literal_size,
+    r = mailimap_literal_data_send(session->imap_stream, literal, (uint32_t)literal_size,
                                    session->imap_progr_rate, session->imap_progr_fun);
   }
   if (r != MAILIMAP_NO_ERROR)
