+++ src/low-level/imap/mailimap_sender.h
@@ -160,7 +160,7 @@ int
 mailimap_literal_count_send(mailstream * fd, uint32_t count);
 
 int
-mailimap_literal_data_send(mailstream * fd, const char * literal, size_t len,
+mailimap_literal_data_send(mailstream * fd, const char * literal, uint32_t len,
 			   size_t progr_rate,
 			   progress_function * progr_fun);
